userDelete function added

+++ src/controller/deleteuser.controller.ts
@@ -4,7 +4,12 @@ const prisma = new PrismaClient();
 
 const deleteUser = async (req: Request, res: Response) => {
   try {
-    const userId = req.params.id;
+    const userId = req.user?.id;
+
+    if (!userId) {
+      res.status(401).json({ message: 'Unauthorized: User not found' });
+      return;
+    }
 
     await prisma.message.deleteMany({
       where: {
@@ -21,7 +26,7 @@ const deleteUser = async (req: Request, res: Response) => {
     await prisma.user.delete({
       where: { id: userId },
     });
-    res.status(200).json({ message: 'user deleted Successfully' });
+    res.status(204).json({ message: 'user deleted Successfully' });
   } catch (error) {
     console.error(error, 'Deleting user Failed');
     res.status(500).json({ message: 'User deletation failed' });
